feat: rewrite connect/disconnect code
