+++ HeartRateLE.Bluetooth/PairingHelper.cs
@@ -18,6 +18,7 @@ namespace HeartRateLE.Bluetooth
             {
                 device.DeviceInformation.Pairing.Custom.PairingRequested += Custom_PairingRequested;
                 var result = await device.DeviceInformation.Pairing.Custom.PairAsync(DevicePairingKinds.ConfirmOnly);
+                //var result = await device.DeviceInformation.Pairing.PairAsync();
 
                 return new Schema.PairingResult()
                 {
